int to long chnages for report

+++ src/main/java/cc/altius/FASP/model/report/ConsumptionForecastVsActualOutputRowMapper.java
@@ -19,11 +19,11 @@ public class ConsumptionForecastVsActualOutputRowMapper implements RowMapper<Con
     public ConsumptionForecastVsActualOutput mapRow(ResultSet rs, int i) throws SQLException {
         ConsumptionForecastVsActualOutput cfa = new ConsumptionForecastVsActualOutput();
         cfa.setTransDate(rs.getDate("MONTH"));
-        cfa.setForecastedConsumption(rs.getInt("FORECASTED_CONSUMPTION"));
+        cfa.setForecastedConsumption(rs.getLong("FORECASTED_CONSUMPTION"));
         if (rs.wasNull()) {
             cfa.setForecastedConsumption(null);
         }
-        cfa.setActualConsumption(rs.getInt("ACTUAL_CONSUMPTION"));
+        cfa.setActualConsumption(rs.getLong("ACTUAL_CONSUMPTION"));
         if (rs.wasNull()) {
             cfa.setActualConsumption(null);
         }
